+++ src/Sulu/Bundle/MediaBundle/Entity/MediaDataProviderRepository.php
@@ -111,6 +111,8 @@ class MediaDataProviderRepository implements DataProviderRepositoryInterface
             ->leftJoin($alias . '.files', 'file')
             ->leftJoin('file.fileVersions', 'fileVersion', 'WITH', 'fileVersion.version = file.version')
             ->leftJoin('fileVersion.tags', 'tag')
+            ->leftJoin('fileVersion.categories', 'categories')
+            ->leftJoin('categories.translations', 'categoryTranslations')
             ->leftJoin('fileVersion.meta', 'fileVersionMeta')
             ->leftJoin('fileVersion.defaultMeta', 'fileVersionDefaultMeta')
             ->leftJoin('fileVersion.contentLanguages', 'fileVersionContentLanguage')
@@ -155,6 +157,14 @@ class MediaDataProviderRepository implements DataProviderRepositoryInterface
             ->innerJoin('file.fileVersions', 'fileVersion', 'WITH', 'fileVersion.version = file.version');
 
         return 'fileVersion.tags';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function appendCategoriesRelation()
+    {
+        return 'fileVersion.categories';
     }
 
     /**