+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/models/media.js
@@ -22,6 +22,7 @@ define(['mvc/relationalmodel'], function (RelationalModel) {
                 name: '',
                 description: '',
                 tags: [],
+                categories: [],
                 copyright: '',
                 credits: '',
                 versions: {}